+++ caf-audit-binding-elasticsearch/src/main/resources/CafAuditEventTenantIndexMappings.json
