+++ caf-audit-binding-elasticsearch/src/main/resources/META-INF/services/com.hpe.caf.auditing.AuditConnectionProvider
