+++ caf-audit-binding-elasticsearch/src/test/java/com/hpe/caf/auditing/AuditConnectionHelper.java
