+++ caf-audit-binding-elasticsearch/src/test/java/com/hpe/caf/auditing/elastic/ElasticAuditIT.java
