+++ caf-audit-binding-webservice/src/main/java/com/hpe/caf/auditing/webserviceclient/WebServiceClientException.java
