+++ caf-audit-binding-webservice/src/test/java/com/hpe/caf/auditing/AuditConnectionHelper.java
