+++ caf-audit-binding-webservice/src/test/java/com/hpe/caf/auditing/webserviceclient/WebServiceClientAuditChannelTest.java
