+++ caf-audit-service-container/pom.xml
@@ -44,11 +44,11 @@
         <!-- Auditting Implemenations -->
         <dependency>
             <groupId>com.github.cafaudit</groupId>
-            <artifactId>caf-audit-elasticclient</artifactId>
+            <artifactId>caf-audit-binding-elasticsearch</artifactId>
         </dependency>
         <dependency>
             <groupId>com.github.cafaudit</groupId>
-            <artifactId>caf-audit-webserviceclient</artifactId>
+            <artifactId>caf-audit-binding-webservice</artifactId>
         </dependency>
         <!-- End of Auditting Implemenations -->
         <dependency>
