--- caf-audit-webserviceclient/src/main/resources/CafAuditEventTenantIndexMappings.json
@@ -1,110 +1,0 @@
-{
-  "cafAuditEvent":{
-    "properties":{
-      "processId":{
-        "type":"keyword"
-      },
-      "threadId":{
-        "type":"long"
-      },
-      "eventOrder":{
-        "type":"long"
-      },
-      "eventTime":{
-        "type":"date"
-      },
-      "eventTimeSource":{
-        "type":"keyword"
-      },
-      "userId":{
-        "type":"keyword"
-      },
-      "correlationId":{
-        "type":"keyword"
-      },
-      "eventCategoryId":{
-        "type":"keyword"
-      },
-      "eventTypeId":{
-        "type":"keyword"
-      },
-      "applicationId":{
-        "type":"keyword"
-      }
-    },
-    "dynamic_templates":[
-      {
-        "CAFAuditKeyword":{
-          "match":"*_CAKyw",
-          "mapping":{
-            "type":"keyword"
-          }
-        }
-      },
-      {
-        "CAFAuditText":{
-          "match":"*_CATxt",
-          "mapping":{
-            "type":"text"
-          }
-        }
-      },
-      {
-        "CAFAuditLong":{
-          "match":"*_CALng",
-          "mapping":{
-            "type":"long"
-          }
-        }
-      },
-      {
-        "CAFAuditInteger":{
-          "match":"*_CAInt",
-          "mapping":{
-            "type":"integer"
-          }
-        }
-      },
-      {
-        "CAFAuditShort":{
-          "match":"*_CAShort",
-          "mapping":{
-            "type":"short"
-          }
-        }
-      },
-      {
-        "CAFAuditDouble":{
-          "match":"*_CADbl",
-          "mapping":{
-            "type":"double"
-          }
-        }
-      },
-      {
-        "CAFAuditFloat":{
-          "match":"*_CAFlt",
-          "mapping":{
-            "type":"float"
-          }
-        }
-      },
-      {
-        "CAFAuditDate":{
-          "match":"*_CADte",
-          "mapping":{
-            "type":"date"
-          }
-        }
-      },
-      {
-        "CAFAuditBoolean":{
-          "match":"*_CABln",
-          "mapping":{
-            "type":"boolean"
-          }
-        }
-      }
-    ]
-  }
-}
