--- caf-audit/test-case/events.xml
@@ -1,138 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright 2015-2018 Micro Focus or one of its affiliates.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-         http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<AuditedApplication
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns="https://cafaudit.github.io/audit-service/schema/AuditedApplication.xsd"
-    xsi:schemaLocation="https://cafaudit.github.io/audit-service/schema/AuditedApplication.xsd"
->
-    <ApplicationId>TestAuditEvents</ApplicationId>
-    <AuditEvents>
-        <AuditEvent>
-            <TypeId>TestEvent1</TypeId>
-            <CategoryId>TestCategory1</CategoryId>
-            <Params>
-                <Param>
-                    <Name>StringType</Name>
-                    <Type>string</Type>
-                    <ColumnName>StringType</ColumnName>
-                    <Description>Description for StringType</Description>
-                </Param>
-                <Param>
-                    <Name>StringKeywordType</Name>
-                    <Type>string</Type>
-                    <IndexingHint>keyword</IndexingHint>
-                    <Description>Description for StringKeywordType</Description>
-                </Param>
-                <Param>
-                    <Name>StringTextType</Name>
-                    <Type>string</Type>
-                    <IndexingHint>fulltext</IndexingHint>
-                    <Description>Description for StringTextType</Description>
-                </Param>
-                <Param>
-                    <Name>StringWithConstraintsType</Name>
-                    <Type>string</Type>
-                    <Constraints>
-                        <MinLength>1</MinLength>
-                        <MaxLength>100</MaxLength>
-                    </Constraints>
-                    <Description>Description for StringType</Description>
-                </Param>
-                <Param>
-                    <Name>ShortType</Name>
-                    <Type>short</Type>
-                    <ColumnName>ShortType</ColumnName>
-                    <Description>Description for Int16Type</Description>
-                </Param>
-                <Param>
-                    <Name>IntType</Name>
-                    <Type>int</Type>
-                    <ColumnName>IntType</ColumnName>
-                    <Description>Description for Int32Type</Description>
-                </Param>
-                <Param>
-                    <Name>LongType</Name>
-                    <Type>long</Type>
-                    <ColumnName>LongType</ColumnName>
-                    <Description>Description for Int64Type</Description>
-                </Param>
-                <Param>
-                    <Name>FloatType</Name>
-                    <Type>float</Type>
-                    <ColumnName>FloatType</ColumnName>
-                    <Description>Description for FloatType</Description>
-                </Param>
-                <Param>
-                    <Name>DoubleType</Name>
-                    <Type>double</Type>
-                    <ColumnName>DoubleType</ColumnName>
-                    <Description>Description for DoubleType</Description>
-                </Param>
-                <Param>
-                    <Name>BooleanType</Name>
-                    <Type>boolean</Type>
-                    <ColumnName>BooleanType</ColumnName>
-                    <Description>Description for BooleanType</Description>
-                </Param>
-                <Param>
-                    <Name>DateType</Name>
-                    <Type>date</Type>
-                    <ColumnName>DateType</ColumnName>
-                    <Description>Description for DateType</Description>
-                </Param>
-            </Params>
-        </AuditEvent>
-        <AuditEvent>
-            <TypeId>TestEvent2</TypeId>
-            <CategoryId>TestCategory2</CategoryId>
-            <Params>
-                <Param>
-                    <Name>StringType2</Name>
-                    <Type>string</Type>
-                    <ColumnName>StringType2</ColumnName>
-                    <Description>Description for StringType2</Description>
-                </Param>
-                <Param>
-                    <Name>ShortType2</Name>
-                    <Type>short</Type>
-                    <ColumnName>ShortType2</ColumnName>
-                    <Description>Description for Int16Type2</Description>
-                </Param>
-            </Params>
-        </AuditEvent>
-        <AuditEvent>
-            <TypeId>TestEvent3</TypeId>
-            <CategoryId>TestCategory3</CategoryId>
-            <Params>
-                <Param>
-                    <Name>StringType3</Name>
-                    <Type>string</Type>
-                    <ColumnName>StringType3</ColumnName>
-                    <Description>Description for StringType3</Description>
-                </Param>
-                <Param>
-                    <Name>DateType2</Name>
-                    <Type>date</Type>
-                    <ColumnName>DateType2</ColumnName>
-                    <Description>Description for DateType2</Description>
-                </Param>
-            </Params>
-        </AuditEvent>
-    </AuditEvents>
-</AuditedApplication>
